requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///D:/bld/aiohttp_1752162621143/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
annotated-types==0.7.0
anyio==4.9.0
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
Brotli==1.1.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_67mh6i36ci/croot/certifi_1752653721535/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
contourpy==1.3.2
cycler==0.12.1
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1746740462479/work
diffusers==0.34.0
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1706434688412/work
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.0
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
fonttools==4.59.0
frozenlist @ file:///D:/bld/frozenlist_1752167168686/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1741403990995/work
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
hf-xet @ file:///D:/bld/bld/rattler-build_hf-xet_1750541236/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1752514897442/work
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==8.7.0
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.3
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
monai-weekly==1.6.dev2528
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict @ file:///D:/bld/multidict_1751310589600/work
multiprocess @ file:///D:/bld/multiprocess_1724954696823/work
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy @ file:///C:/b/abs_14sk27cuu6/croot/numpy_and_numpy_base_1747238041220/work/dist/numpy-2.2.5-cp310-cp310-win_amd64.whl#sha256=5289707c9475225bacfd6f4358af6149cf7f686595330aef5bf2198e61e5d36f
opencv-python==4.12.0.88
orjson==3.11.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///D:/bld/pandas_1752081728932/work
pillow @ file:///C:/b/abs_b06nvnir5m/croot/pillow_1752524543576/work
propcache @ file:///D:/bld/propcache_1744524966467/work
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow @ file:///C:/b/abs_0aneze9pgl/croot/pyarrow_1738229440651/work/python
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-multipart==0.0.20
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex @ file:///D:/bld/regex_1730952192742/work
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich==14.0.0
ruff==0.12.4
safehttpx==0.1.6
safetensors @ file:///D:/bld/safetensors_1740651461615/work
scipy @ file:///C:/b/abs_71q96u23a4/croot/scipy_1747238057723/work/dist/scipy-1.15.3-cp310-cp310-win_amd64.whl#sha256=06013d45c5cad262bccdca1fbe5c25a720be04ae66dc83d75d1f749d118a7346
semantic-version==2.10.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
starlette==0.47.1
sympy @ file:///C:/b/abs_b4u17p23yg/croot/sympy_1738108511395/work
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1752247352056/work
typer==0.16.0
typing-inspection==0.4.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
uvicorn==0.35.0
websockets==15.0.1
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
xxhash @ file:///D:/bld/python-xxhash_1740594804140/work
yarl @ file:///D:/bld/yarl_1749554852283/work
zipp==3.23.0
